% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{aplos_download_results}
\alias{aplos_download_results}
\title{Download Results Information from Aplos NCA Analysis}
\usage{
aplos_download_results(url, token, execution_id)
}
\arguments{
\item{url}{The base API URL.}

\item{token}{JWT token.}

\item{execution_id}{The execution ID.}
}
\value{
A data frame with download URL and filename.
}
\description{
Retrieves download URLs and files from Aplos NCA analysis results.
}
\examples{
\dontrun{
  df <- aplos_download_results("https://api.app.aplos-nca.com", token, "exec_123")
}
}
