% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{aplos_execute_analysis}
\alias{aplos_execute_analysis}
\title{Execute NCA Analysis in Aplos}
\usage{
aplos_execute_analysis(
  result,
  data_cleaning = "{}",
  analysis = "{}",
  calcs = "{}",
  plots = "{}",
  tables = "{}",
  metadata = "{}",
  url,
  token,
  name = "NCA Analysis via R",
  save_body = FALSE
)
}
\arguments{
\item{result}{Result from aplos_get_upload_url().}

\item{data_cleaning}{JSON string for data cleaning (default "{}").}

\item{analysis}{JSON string for PK parameter calculation (default "{}").}

\item{calcs}{JSON string for custom calculations (default "{}").}

\item{plots}{JSON string for custom plots (default "{}").}

\item{tables}{JSON string for custom tables (default "{}").}

\item{metadata}{JSON string for metadata (default "{}").}

\item{url}{The base API URL.}

\item{token}{JWT token.}

\item{name}{Analysis name (default "NCA Analysis via R").}

\item{save_body}{Logical; save body of API POST command to body.txt file? (default FALSE, set to TRUE for debugging).}
}
\value{
The execution ID.
}
\description{
Submits and executes an NCA analysis on uploaded data.
}
\examples{
\dontrun{
  exec_id <- aplos_execute_analysis(upload_result,
  url = "https://api.app.aplos-nca.com", token = token)
}
}
