% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{aplos_get_jwt}
\alias{aplos_get_jwt}
\title{Obtain Aplos NCA API JWT Token}
\usage{
aplos_get_jwt(client_id, username, password, region)
}
\arguments{
\item{client_id}{The Cognito client ID.}

\item{username}{Your Aplos NCA username.}

\item{password}{Your Aplos NCA password.}

\item{region}{The AWS region (e.g., "us-east-1").}
}
\value{
A character string containing the JWT token.
}
\description{
Authenticates with the Aplos NCA API using AWS Cognito. This function sends a POST request to retrieve a JWT token for subsequent API calls. All API calls require a valid JWT. The JWT is only valid for a period of time (~15 minutes). Each time you make an API call with a valid JWT, that token is refreshed and the valid period is extended. Thus during a normal analysis you will only need to authenticate once as you will be making API calls within the valid time window. However, if you walk away from your computer or there is a long delay between API calls you may need to authenticate again.

Note: Store credentials securely; do not hardcode them in scripts. It is recommended that you use a hidden text file (e.g. ".aplos_creds") stored on your local computer. Read that file in and then parse it to update the required information for authentication. An example credential file may look like the following:
APLOS_API_URL="https://api.app.aplos-nca.com"
COGNITO_CLIENT_KEY="your_cognito_client_id/key"
COGNITO_USER_NAME="your_username"
COGNITO_PASSWORD="your_password"
COGNITO_REGION="your_aws_region"
All of this information can be found in your user profile for your Aplos NCA account under "API Settings".
}
\examples{
\dontrun{
  token <- aplos_get_jwt(client_id = "your_client_id", username = "user",
                         password = "pass", region = "us-east-1")
}
}
