% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{aplos_upload_file}
\alias{aplos_upload_file}
\title{Upload File to Aplos NCA}
\usage{
aplos_upload_file(input_file, result, token)
}
\arguments{
\item{input_file}{Path to the input file.}

\item{result}{Result from aplos_get_upload_url().}

\item{token}{JWT token.}
}
\description{
Uploads a file using the presigned URL.
}
\examples{
\dontrun{
  aplos_upload_file("path/to/file.csv", upload_result, token)
}
}
